+++ kreate-plugin/src/main/kotlin/com/davils/kreate/module/project/ProjectModule.kt
@@ -19,6 +19,7 @@ package com.davils.kreate.module.project
 import com.davils.kreate.KreateExtension
 import com.davils.kreate.module.Module
 import com.davils.kreate.module.project.api.initializeApiValidation
+import com.davils.kreate.module.project.benchmark.initializeBenchmark
 import com.davils.kreate.module.project.constants.initializeBuildConstants
 import com.davils.kreate.module.project.detekt.initializeDetekt
 import com.davils.kreate.module.project.docs.initializeDocs
@@ -70,6 +71,7 @@ internal object ProjectModule : Module {
             initializePublish(extension)
             initializeDetekt(extension)
             initializeApiValidation(extension)
+            initializeBenchmark(extension)
         }
     }
 }